aml-elliptic-gnn/requirements.txt
matplotlib>=3.7.1
numpy>=1.22.4
pandas==1.5.3
PyYAML>=6.0
scikit_learn>=1.2.2
torch>=2.0.0+cu118
torch_geometric>=2.3.1
requests
